Add new splash screen and remove unused banner files - fixes #1716 (#1803)
* add new splash screen as decided in #1716

* tweak files to avoid edge artifact

* remove unused banner files

+++ raster/r.surf.random/local_proto.h
@@ -1,2 +1,2 @@
 /* randsurf.c */
-int randsurf(char *, double, double, int);
+int randsurf(char *, int, int, int);
